Simplify GET:/sever and throw a 'version' on the path.
