+++ src/handlers/server/index.mjs
@@ -2,11 +2,11 @@ import { handlers } from './errors'
 import { handlers as pluginHandlers } from './plugins'
 
 import * as api from './api'
-import * as get from './get'
 import * as nextCommands from './next-commands'
 import * as reload from './reload'
 import * as stop from './stop'
+import * as version from './version'
 
-handlers.push(api, get, nextCommands, reload, stop, ...pluginHandlers)
+handlers.push(api, nextCommands, reload, stop, version, ...pluginHandlers)
 
 export { handlers }
